changes in cccp and added 2 methods for db wirting in support

+++ Biglietteria/src/it/unipv/po/cccp/pagamento/CalcolatoreSconto.java
@@ -1,12 +1,11 @@
 package it.unipv.po.cccp.pagamento;
 
-import java.time.LocalDateTime;
-import java.time.Month;
+
 
 public class CalcolatoreSconto implements ICalcolatoreSconto{
 
 	@Override
-	public double calcolaSconto(double totale, int punti) {
+	public double calcolaSconto(double totale, double punti) {
 		
 		return (punti/100)*totale;
 	}
