+++ Biglietteria/src/it/unipv/po/cccp/pagamento/ICalcolatoreSconto.java
@@ -4,7 +4,7 @@ import java.time.*;
 //facilita la futura variazione delle regole 
 //di sconto
 public interface ICalcolatoreSconto {
-		public double calcolaSconto(double totale, int punti);
+		public double calcolaSconto(double totale, double punti);
 
 		
 }
